# bibliotheque-api/openapi.yaml
openapi: '3.0.3'
info:
  title: API Bibliothèque
  version: '2.0.0'
  description: |
    API REST de gestion d'une bibliothèque — Cours 2026.
    Gestion des auteurs, des livres et des emprunts avec authentification JWT.
    **Séance 6** : intégration OpenAI (résumés, recherche intelligente) et Stripe (abonnements premium).
  contact:
    name: Étudiant L3
servers:
  - url: http://localhost:8000/api/v1
    description: Serveur de développement
  - url: https://api.mabibliotheque.com/v1
    description: Serveur de production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaymentIntentResponse:
      type: object
      description: Réponse de création d'un PaymentIntent Stripe
      properties:
        client_secret:
          type: string
          example: pi_3abc...._secret_xyz
          description: Secret à utiliser côté client avec Stripe.js pour finaliser le paiement
        payment_intent_id:
          type: string
          example: pi_3abc....
        amount:
          type: integer
          example: 500
          description: Montant en centimes (500 = 5,00 €)
        currency:
          type: string
          example: eur
        publishable_key:
          type: string
          example: pk_test_...
          description: Clé publique Stripe à utiliser côté client
    SubscriptionStatusResponse:
      type: object
      description: Statut d'abonnement de l'utilisateur connecté
      properties:
        user_id: { type: integer, example: 1 }
        email: { type: string, example: admin@biblio.local }
        subscription_status:
          type: string
          enum: [free, premium]
          example: free
        borrow_limit:
          type: integer
          nullable: true
          example: 2
          description: "Nombre max d'emprunts simultanés (null = illimité pour premium)"
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: Validation échouée }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
  responses:
    Unauthorized:
      description: Non authentifié — token manquant ou invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
paths:
...
  /subscriptions/payment-intent:
    post:
      tags: [Subscriptions]
      summary: Créer un PaymentIntent Stripe (abonnement premium)
      description: |
        Crée un PaymentIntent Stripe de **5,00 € (500 centimes)** pour passer à l'abonnement premium.
        Retourne un `client_secret` à utiliser côté client avec **Stripe.js** pour finaliser le paiement.
        Le statut premium est activé automatiquement via le **webhook** `payment_intent.succeeded`.
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: PaymentIntent créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/PaymentIntentResponse' }
                  message: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: Utilisateur déjà premium
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: Service Stripe indisponible
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /subscriptions/status:
    get:
      tags: [Subscriptions]
      summary: Statut d'abonnement de l'utilisateur connecté
      description: |
        Retourne le statut d'abonnement (`free` ou `premium`) et la limite d'emprunts.
        - **free** : 2 emprunts simultanés maximum
        - **premium** : emprunts illimités (`borrow_limit: null`)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Statut retourné
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/SubscriptionStatusResponse' }
                  message: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
...